refactor: enforce AGENTS.md SSoT compliance and restructure codebase
- Replace all inline CubicBezierEasing duplicates with EasingTokens.Default references
- Replace all durationMs = 120/220 magic numbers with DurationTokens.Default.shortMs/mediumMs
- Add shortMs/mediumMs/longMs computed properties to DurationTokens
- Remove dead code (InternalEasings.kt)
- Restructure composeApp screens with proper file splitting and state management
- Add new module files: CanimationNamespacesCore/Motion/Style, PresetRegistryDefaultsPart01-03
- Update docs and build configuration

+++ composeApp/src/commonMain/kotlin/com/segnities007/canimation/screen/examples/ComponentAnimations6.kt
@@ -305,7 +305,7 @@ fun WaveProgressBar(modifier: Modifier = Modifier) {
         drawRoundRect(color = trackColor, size = size, cornerRadius = androidx.compose.ui.geometry.CornerRadius(h / 2))
         val filledW = w * progress
         for (x in 0..filledW.toInt()) {
-            val waveY = (sin(phase + x * 0.05f) * 2).toFloat()
+            val waveY = sin(phase + x * 0.05f) * 2
             drawCircle(color = primaryColor, radius = 1.5f, center = androidx.compose.ui.geometry.Offset(x.toFloat(), h / 2 + waveY))
         }
     }
